Update status code for errors

+++ delete_Org.js
@@ -78,9 +78,16 @@ router.delete('/', async (req, res) => {
             return res.status(400).json({ error: 'Organization ID is required!' });
         }
 
-        // Execute SQL query to delete organization
-        const result = await pool.query('DELETE FROM organization WHERE org_ID = ?', [org_ID]);
+        // Check if there are users associated with the organization
+        const userResult = await pool.query('SELECT COUNT(*) AS userCount FROM users WHERE Org_ID = ?', [org_ID]);
+        const userCount = userResult[0][0].userCount;
 
+        if (userCount > 0 ) {
+            return res.status(300).json({ error: 'Cannot delete organization as it is associated with users' });
+        }else{ 
+
+             const result = await pool.query('DELETE FROM organization WHERE org_ID = ?', [org_ID]);
+       
         // Check if any rows were affected
         if (result.affectedRows === 0) {
             return res.status(404).json({ error: 'Organization not found!' });
@@ -88,8 +95,11 @@ router.delete('/', async (req, res) => {
 
         // Send success response
         return res.status(200).json({ message: 'Organization deleted successfully' });
+    }
     } catch (error) {
         console.error("Error deleting organization:", error);
+        
+        // Handle other errors
         return res.status(500).json({ error: 'Error deleting organization' });
     }
 });
